+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Added ability to control+click and shift+click select multiple folders in the sidebar. This allows you to filter and view only the articles contained within those folders.
 
+#### Show folder feed count in sidebar
+
+- Added "Show folder feed count" toggle to Sidebar settings tab (off by default) which tallies up the amount of feeds contained with a parent folder (includes subfolder feeds)
+
 ## 2.4.0-beta.3 - June 17, 2026
 
 ### Features